feat: add preview profile modal and adjust modal layout in profile page
