--- src/pages/profile/preview-profile/index.js
@@ -1,15 +1,0 @@
-import { NavBar } from "@/components/NavBar";
-import { PreviewProfile } from "@/components/PreviewProfile";
-
-export default function PreviewProfilePage() {
-  return (
-    <div>
-      <nav className="nav-bar-section">
-        <NavBar />
-      </nav>
-      <main className="preview-section">
-        <PreviewProfile />
-      </main>
-    </div>
-  );
-}
